Avoid initializing the playback proxy for a few remaining tests
This list was obtained by temporarily causing playback to panic if
there was no `http-data` file when the proxy is started.

The only remaining test that starts a proxy server which goes unused is
`krate::yank_not_owner`, however this test does need an index.  If more
tests fall into this category over time we can switch to a full builder
API for orchestrating a `TestApp` but that doesn't seem necessary yet.