Define `recent_crate_downloads` in terms of `version_downloads`
The only place we're ever using `crate_downloads` directly is in a place
that we should have been using `recent_crate_downloads`. Once we make
that change, `crate_downloads` is only ever used to populate
`recent_crate_downloads`, which can just as easily be populated from
`version_downloads` instead.

Populating it from `version_downloads` is a bit slower since the table
is so much larger, and we don't have an index on it which we can use for
this query. However, this doesn't really matter since we run this
asynchronously. Now that we're on PG 11, I also intend to partition
`version_downloads` on date which will make populating the view much
faster once again.

Note that I have not removed `crate_downloads` yet. That needs to be
deployed separately, as it's a destructive change which would break
running code, so we can't run that migration until the code which is no
longer relying on `crate_downloads` is deployed