Rerun the build script if `.env` has been changed.
Edits to `.env` may alter the value of the `TEST_DATABASE_URL` environment variable, so we should
rerun the build script if `.env` was changed. The build script is already configured to be rerun
when the value of `TEST_DATABASE_URL` changes, but the current setup does not capture changes
performed via `.env`, since the environment file is only loaded at runtime of the build script.

This change also removes the redundant `cargo:rerun-if-changed=build.rs`; the build script is always
rerun if it is changed.