Add a working "matches all keywords" query.
Diesel has a bug where array expression methods aren't implemented on
nullable expressions, so we can't just call Diesel's `.contains` method
until the fix lands. I've added a workaround for the time being. This
results in a runtime error since `keyword` was defined as `varchar`,
which coerces to `text`, but arrays do not. I've changed the column type
to address this. The keywords table is small enough and read
infrequently enough that I'm not concerned about the exclusive lock in
this migration.