Adjust branding to reflect what crates.io actually does
To be clear that crates.io is a Rust package registry---in fact the
_default_ crate host for Rust---this commit changes the name,
short_name, and description fields in the manifest to reflect the nature
of the project.  crates.io isn't "Cargo", so it seems logical to me to
not mention much about it, at least in the metadata for the site.

I got the branding "Rust Package Registry" from the tagline in the
visible top bar on the landing page.

Also, update the various acceptance tests that I found floating around.

Signed-off-by: Kristofer Rye <kristofer.rye@gmail.com>