Allow blocking arbitrary traffic without redeploying
This generalizes the `block_ips` module to allow checking against any
header instead of just the `X-Real-Ip` header. This does not enable more
complex matching such as regexes, but it does let us apply the simple
string matching we have today to any header.

Additionally, which headers we check against is now configured by an
environment variable, which is a comma separated list in the form
`Header=VALUE_ENV_VAR`. So to keep the behavior we had today, one would
set `BLOCKED_TRAFFIC="X-Real-Ip=BLOCKED_IPS"`.

This means that when we see traffic patterns that we need to block which
can be handled by the existing logic, we can now begin applying it to
patterns we didn't previosuly anticipate without requiring a redeploy