Don't consider versions with build metadata to be unstable
Our backend was behaving properly in this regard, but the front end was
considering any version that had build metadata to be unstable. There's
no reason for us to have this detection be hand-rolled, we can use a
proper semver library to determine whether something is a prerelease or
not.