Fixed version suggestion logic.
Previously, if it had search everything and found nothing, that means it also cannot find `maxVersion`.
Thus, this commit's solution is to simply find anything that is not yet yanked even if it is unstable.
If even this fails, then we truly have nothing to offer thereby offering an empty page.