use deploy keys for git operations
At the moment crates.io relies on Personal Access Tokens to be able to
push changes to the index, but this is not great from a security point
of view as it's not possible to scope a token to a single repository.

That means if an attacker managed to compromise the server hosting
crates.io they'd be able to compromise other, unrelated parts of the
Rust project infrastructure.

This commit changes crates.io to optionally support using deploy keys
instead of tokens to push new commits to GitHub. Deploy keys are scoped
to a single repository and can be either read-only or read-write (in our
case we need write access).

The key must be stored in the GIT_SSH_KEY environment variable,
base64-encoded. It has to be encoded as that way it's easier to
copy-paste it, and the setup-deploy-keys tool in rust-lang/simpleinfra
generates them that way.