prioritize ssh credentials instead of erroring
Before this commit, if both SSH and HTTP git credentials were present
the background worker would error out, but that behavior would make the
initial deployment harder, as there wouldn't be an easy way to rollback
without also setting the HTTP credentials again.

This commit changes that behavior by defaulting to SSH when both
credentials are present, printing a warning to alert the user of that.