fix username not being provided when pushing to the index with ssh
Git2 uses callbacks in a different way when SSH is involved: in some
cases the first call requests just the username, and then a second call
requests the SSH key. This commit changes the code to accomodate this
behavior. More details on the approach are in the added code comments.