Add Cargo.toml copy button to crate rows
To copy the most recent version of a crate to the clipboard, users would
have to go to the crate's individual page. However, it would probably be
convenient to have the same control on the search results page. To
accomplish this, a new component was created to encapsulate the copy
control and its notifications (success/failure).

Fixes #1587