Bump to latest release of conduit-hyper
This upstream release now percent decodes the path component but not the
query string. This behavior aligns with the `civet` server. The known
difference is that `conduit-hyper` does a lossy utf8 conversion while
`civet` panics on invalid utf8 and closes the connection immediately.

This seems like a good compromise of matching the existing behavior as
closely as possible without copying the panic. I looked into fixing the
panic in `civet`, however the `to_str_slice` function returns an
`Option<&str>` and there is nowhere to store a newly allocated `String`
so changing the behavior there is not practical.

https://github.com/jtgeibel/conduit-hyper/compare/v0.3.0-alpha.3...v0.3.0-alpha.4