Header: Replace "Login" link with button
Having this as a link suggests that users e.g. can open this route in a new tab or save a bookmark, but the fact that this is currently implemented as a route should be considered an internal implementation detail and opening the route directly doesn't actually work at all.