CrateOwnerInvite: Avoid "missing `id`" issue
Ember Data expects all resources to have an `id`. In this serializer we are renaming that expectation to `crate_id` for questionable reasons. This worked fine as long as only `CrateOwnerInvite` resources were contained in the response, but now we also have `User` resources, and those don't have a `crate_id` attribute, which is causing issues.

This commit removes the `users` array from the response for now to fix this problem. In the future we should look into either responding with a proper ID or generating it based on `crate_id` and `invitee_id`.