split logged instance metrics into multiple lines when necessary
Heroku has a limit for log lines of 10,000 bytes, and when a line
reaches that size the platform will automatically split it into multiple
lines, breaking the tooling reading it.

To avoid that problem when outputting metrics, this commit implements
the splitting on the application side when a line is longer than 5,000
bytes. The threshold is just 5,000 because the newline is inserted
*after* the item that made the line too long, so it's possible that the
actual lines will be longer than 5,000 bytes.