render: add HTML image tests with align support
The current align test was only used for text, but we want to be sure
that images can be aligned too. This is why a new test is added to
verify that the align attribute is also kept for images.

Since the align attribute requires RAW HTML for the <img> tag, the
relative_links test has also been updated to check for this usage.